break: extract `format` and `cleanNumber` method form Number to `components/number/NumberUtils` file
