+++ packages/dnb-design-system-portal/src/docs/uilib/about-the-lib/releases/v9-info.md
@@ -1,7 +1,3 @@
----
-draft: true
----
-
 # v9
 
 - [Migration](#migration)
@@ -12,6 +8,7 @@ draft: true
 v9 contains a couple of _braking changes_. As a migration process, you can simply search and replace:
 
 1. Find `formRow` inside a (e.g. <Provider formRow= ...) and replace it with `FormRow` (PascalCase).
+1. In case you use `format` or `cleanNumber` from either `/components/Number` or `/components/number/Number`, replace the import path to be `/components/number/NumberUtils`.
 
 ## Changes
